backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.31
alembic==1.13.1
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
supabase==2.4.3
pg8000==1.30.5
python-dotenv==1.0.1
qrcode[pil]==7.4.2
Pillow==10.4.0
email-validator==2.1.1
pymysql==1.1.1
cryptography==42.0.8
apscheduler==3.10.4
pandas==2.2.2
